Split lib into module files
It's getting a little crowded in here.
