+++ src/node-part.ts
@@ -1,0 +1,171 @@
+/**
+ * @license
+ * Copyright (c) 2017 The Polymer Project Authors. All rights reserved.
+ * This code may only be used under the BSD style license found at
+ * http://polymer.github.io/LICENSE.txt
+ * The complete set of authors may be found at
+ * http://polymer.github.io/AUTHORS.txt
+ * The complete set of contributors may be found at
+ * http://polymer.github.io/CONTRIBUTORS.txt
+ * Code distributed by Google as part of the polymer project is also
+ * subject to an additional IP rights grant found at
+ * http://polymer.github.io/PATENTS.txt
+ */
+
+import {getValue} from './lit-html.js';
+import {removeNodes} from './nodes.js';
+import {SinglePart} from './part.js';
+import {TemplateInstance} from './template-instance.js';
+import {TemplateResult} from './template-result.js';
+
+export class NodePart implements SinglePart {
+  instance: TemplateInstance;
+  startNode: Node;
+  endNode: Node;
+  private _previousValue: any;
+
+  constructor(instance: TemplateInstance, startNode: Node, endNode: Node) {
+    this.instance = instance;
+    this.startNode = startNode;
+    this.endNode = endNode;
+    this._previousValue = undefined;
+  }
+
+  setValue(value: any): void {
+    value = getValue(this, value);
+
+    if (value === null ||
+        !(typeof value === 'object' || typeof value === 'function')) {
+      // Handle primitive values
+      // If the value didn't change, do nothing
+      if (value === this._previousValue) {
+        return;
+      }
+      this._setText(value);
+    } else if (value instanceof TemplateResult) {
+      this._setTemplateResult(value);
+    } else if (Array.isArray(value) || value[Symbol.iterator]) {
+      this._setIterable(value);
+    } else if (value instanceof Node) {
+      this._setNode(value);
+    } else if (value.then !== undefined) {
+      this._setPromise(value);
+    } else {
+      // Fallback, will render the string representation
+      this._setText(value);
+    }
+  }
+
+  private _insert(node: Node) {
+    this.endNode.parentNode!.insertBefore(node, this.endNode);
+  }
+
+  private _setNode(value: Node): void {
+    if (this._previousValue === value) {
+      return;
+    }
+    this.clear();
+    this._insert(value);
+    this._previousValue = value;
+  }
+
+  private _setText(value: string): void {
+    const node = this.startNode.nextSibling!;
+    if (node === this.endNode.previousSibling &&
+        node.nodeType === Node.TEXT_NODE) {
+      // If we only have a single text node between the markers, we can just
+      // set its value, rather than replacing it.
+      // TODO(justinfagnani): Can we just check if _previousValue is
+      // primitive?
+      node.textContent = value;
+    } else {
+      this._setNode(document.createTextNode(value === undefined ? '' : value));
+    }
+    this._previousValue = value;
+  }
+
+  private _setTemplateResult(value: TemplateResult): void {
+    let instance: TemplateInstance;
+    if (this._previousValue &&
+        this._previousValue.template === value.template) {
+      instance = this._previousValue;
+    } else {
+      instance =
+          new TemplateInstance(value.template, this.instance._partCallback);
+      this._setNode(instance._clone());
+      this._previousValue = instance;
+    }
+    instance.update(value.values);
+  }
+
+  private _setIterable(value: any): void {
+    // For an Iterable, we create a new InstancePart per item, then set its
+    // value to the item. This is a little bit of overhead for every item in
+    // an Iterable, but it lets us recurse easily and efficiently update Arrays
+    // of TemplateResults that will be commonly returned from expressions like:
+    // array.map((i) => html`${i}`), by reusing existing TemplateInstances.
+
+    // If _previousValue is an array, then the previous render was of an
+    // iterable and _previousValue will contain the NodeParts from the previous
+    // render. If _previousValue is not an array, clear this part and make a new
+    // array for NodeParts.
+    if (!Array.isArray(this._previousValue)) {
+      this.clear();
+      this._previousValue = [];
+    }
+
+    // Lets us keep track of how many items we stamped so we can clear leftover
+    // items from a previous render
+    const itemParts = this._previousValue as any[];
+    let partIndex = 0;
+
+    for (const item of value) {
+      // Try to reuse an existing part
+      let itemPart = itemParts[partIndex];
+
+      // If no existing part, create a new one
+      if (itemPart === undefined) {
+        // If we're creating the first item part, it's startNode should be the
+        // container's startNode
+        let itemStart = this.startNode;
+
+        // If we're not creating the first part, create a new separator marker
+        // node, and fix up the previous part's endNode to point to it
+        if (partIndex > 0) {
+          const previousPart = itemParts[partIndex - 1];
+          itemStart = previousPart.endNode = document.createTextNode('');
+          this._insert(itemStart);
+        }
+        itemPart = new NodePart(this.instance, itemStart, this.endNode);
+        itemParts.push(itemPart);
+      }
+      itemPart.setValue(item);
+      partIndex++;
+    }
+
+    if (partIndex === 0) {
+      this.clear();
+      this._previousValue = undefined;
+    } else if (partIndex < itemParts.length) {
+      const lastPart = itemParts[partIndex - 1];
+      // Truncate the parts array so _previousValue reflects the current state
+      itemParts.length = partIndex;
+      this.clear(lastPart.endNode.previousSibling!);
+      lastPart.endNode = this.endNode;
+    }
+  }
+
+  private _setPromise(value: Promise<any>): void {
+    this._previousValue = value;
+    value.then((v: any) => {
+      if (this._previousValue === value) {
+        this.setValue(v);
+      }
+    });
+  }
+
+  clear(startNode: Node = this.startNode) {
+    removeNodes(
+        this.startNode.parentNode!, startNode.nextSibling!, this.endNode);
+  }
+}
