+++ src/part.ts
@@ -1,0 +1,29 @@
+/**
+ * @license
+ * Copyright (c) 2017 The Polymer Project Authors. All rights reserved.
+ * This code may only be used under the BSD style license found at
+ * http://polymer.github.io/LICENSE.txt
+ * The complete set of authors may be found at
+ * http://polymer.github.io/AUTHORS.txt
+ * The complete set of contributors may be found at
+ * http://polymer.github.io/CONTRIBUTORS.txt
+ * Code distributed by Google as part of the polymer project is also
+ * subject to an additional IP rights grant found at
+ * http://polymer.github.io/PATENTS.txt
+ */
+
+import {TemplateInstance} from './template-instance.js';
+
+export interface Part {
+  instance: TemplateInstance;
+  size?: number;
+}
+
+export interface SinglePart extends Part {
+  setValue(value: any): void;
+}
+
+export interface MultiPart extends Part {
+  setValue(values: any[], startIndex: number): void;
+}
+
